+++ src/Modules/Admin/Controllers/Traits/LocationTrait.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Hirtz\Location\Modules\Admin\Controllers\Traits;
 
 use Hirtz\Location\Models\Location;
-use Yii;
 use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 
@@ -28,7 +27,7 @@ trait LocationTrait
 
     protected function checkLocationPermission($location, $permissionName): void
     {
-        if (!Yii::$app->getUser()->can($permissionName, ['location' => $location])) {
+        if (!$this->webuser->can($permissionName, ['location' => $location])) {
             throw new ForbiddenHttpException();
         }
     }