+++ src/Modules/Admin/Module.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Hirtz\Location\Modules\Admin;
 
-use Hirtz\Skeleton\I18n\Lang;
 use Hirtz\Cms\Models\Entry;
 use Hirtz\Location\Models\Location;
 use Hirtz\Location\Modules\Admin\Interfaces\AutocompleteInterface;
@@ -35,7 +34,7 @@ class Module extends \Hirtz\Skeleton\Base\Module implements ModuleInterface
     {
         return $dashboard->addItem(DashboardItem::make()
             ->icon('map-marker-alt')
-            ->label(Lang::t('location', 'LOCATION_CREATE_BUTTON'))
+            ->label(Yii::t('location', 'LOCATION_CREATE_BUTTON'))
             ->roles([Location::AUTH_LOCATION_CREATE])
             ->url(['/admin/location/location/create']));
     }