+++ src/Modules/Admin/Widgets/Forms/LocationProviderIdField.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Hirtz\Location\Modules\Admin\Widgets\Forms;
 
-use Hirtz\Skeleton\I18n\Lang;
 use Hirtz\Location\Modules\Admin\Assets\AutocompleteAssetBundle;
 use Hirtz\Location\Modules\Admin\Module;
 use Hirtz\Location\Modules\ModuleTrait;
@@ -24,7 +23,7 @@ class LocationProviderIdField extends InputField
         $this->attributes['autocomplete'] ??= 'off';
 
         if (!$this->model->{$this->property}) {
-            $this->attributes['placeholder'] ??= Lang::t('location', 'LOCATION_PROVIDER_ID_SEARCH');
+            $this->attributes['placeholder'] ??= Yii::t('location', 'LOCATION_PROVIDER_ID_SEARCH');
             $this->attributes['type'] ??= 'search';
         }