+++ src/Modules/Admin/Widgets/Navs/LocationSubmenu.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Hirtz\Location\Modules\Admin\Widgets\Navs;
 
-use Hirtz\Skeleton\I18n\Lang;
 use Hirtz\Location\Models\Location;
 use Hirtz\Location\Modules\ModuleTrait;
 use Hirtz\Skeleton\Widgets\Navs\NavItem;
@@ -32,7 +31,7 @@ class LocationSubmenu extends Submenu
     protected function getLocationItem(): ?NavItem
     {
         return NavItem::make()
-            ->label(Lang::t('skeleton', 'COMMON_GENERAL'))
+            ->label(Yii::t('skeleton', 'COMMON_GENERAL'))
             ->url($this->model->getAdminRoute())
             ->routes(['admin/location/location/'])
             ->icon('cog');
@@ -41,7 +40,7 @@ class LocationSubmenu extends Submenu
     protected function getLocationTagsItem(): ?NavItem
     {
         return NavItem::make()
-            ->label(Lang::t('location', 'COMMON_TAGS'))
+            ->label(Yii::t('location', 'COMMON_TAGS'))
             ->url(['location-tag/index', 'location' => $this->model->id])
             ->visible(static::getModule()->enableTags)
             ->badge($this->model->tag_count ?: null)